Remove validation on apply-auto promos with codes
This is important for backward compatibility since if we keep validating
against promotions with codes as well, it will be impossible to update
existing apply-auto promotions that already have a code associated.

Since promotion codes are created in a second step, the creation validation
have to be done in that step.