Move the reimbursement category logic to Spree::Reimbursement
Add Spree::Reimbursement.store_credit_category_proc as a simple way
to redefine how the reimbursement category should be fetched.

Add Spree::Reimbursement#store_credit_category as the most natural
way to fetch a given category for a reimbursement.

This fixes reimbursement_category_name ending up in a translation
missing error. We need to lazily load the translation because, when
the file is loaded, Rails will still have to load translations from
the main application.

Co-Authored-By: Filippo Liverani <dev@mailvore.com>