Remove code from Spree::Api::PromotionsController
Make it leaner by using the way solidus api has to deal with ActiveRecord::RecordNotFound
and authorization.

In `load_promotion`, `find_by` is replaced with `find` and the order of
the finders is inverted to have the exception as last operator: if a
promotion can't be found, raise an exception.