Allow the installer to add solidus_auth_devise
There's no need to manually add the gem to the Gemfile.

Rails installer has a `gem` method that does exactly that,
making the installation process easier for new users.

During the first installation it will be asked to users if
they want to install the gem, still making this optional.

This commit also updates Readme, Guides and avoid installing the
plugin in sandbox and rails application template (where it's already
done manually), and when building the test_app (where we don't need it).