API uploads images via URL
This enhancement give the images API the ability to create and update images from
URLs.  A simple check determines whether or not the attachment parameter is a URL.

If it is a URL: OpenURI is used to download the image and it subsequently
undergoes the standard post upload processing (via paperclip).  After
processing the image, the other attributes (besides :attachment) are updated on
the resultant Spree::Image.

If it is not a URL: the attachment parameter is handled in the same manner
it was prior to this commit.

API documentation added.

Uses static image from solidus repo, to avoid external dependency