Fix api checkout flow documentation
The `Shipment` resource is only accessible by admin user and cannot be directly
accessed by a user with no admin role.

The shipments and related rates are returned in the orders json.
To access the order shipments, users with no admin role should GET
`orders/:order_number` or `/orders/current` and fetch the shipment rates attributes.