Add a preference to determine if redirect back when unauthorized
With this preference we'll be able to communicate with store
owners that the behavior is changing and they have to manually
change the preference and be sure that their store is still
working as expected.

New app will have the new behavior, with the preference set
to true, while existing app will keep having the old behavior
(false), so they can see the deprecation warning.