Rescue & redirect nil-resource links consistently
The errors caused by non-existent resource links were
created at the point of loading the resources. The primary of which was
due to the resource_controller#parent method returning a `nil` when the
parent_id was not found. The solution was to change `find_by` to
`find_by!`, thereby throwing a `RecordNotFound` error, which could be
handled appropriately.

The secondary point of failure (for some) was created at the
resource_controller#collection method, which raised a `NoMethodError` by
attempting to call `parent.send(controller_name)` when `parent` was
`nil`. The expected behavior was that the `RecordNotFound` would be
raised in #parent and the code execution stopped after the successful
error handling. However, due to the behavior of the in-method rescue
(needed for error-handling context) in contrast to `Rescue_from`, the
code continues executing the `.send(controller_name)` after the
redirection, which then results in the mentioned `NoMethodError`. The
solution was to add a second condition in the return statement to ensure
that 'parent' was not 'nil' before `.send(controller_name)` is called.

Additionally, the resource_controller#resource_not_found method was
moved to the parent class (the base_controller). The move allowed all
children of the base_controller to have access to the method for flash
and redirection. The explicit return of a falsy value, `nil`, at the
end of this method stops the code from continuing to run and eventually
causing a 'DoubleRenderError' after the redirect.

The #resource_not_found was overridden in the resource_controller and
allowed to accept optional parameters for cases where the previously
existing redirection was not appropriate or threw errors due to the
incorrect URLs generated by the polymorphic
resource_controller#collection_url method.

Other controllers needed to be rescued on a case by case approach when
loading the resource to have the appropriate context to
redirect and flash the correct messages.

Tests were created for these problematic URLs to ensure consistency in
the future.