Use CircleCI contexts in jobs that require secrets
This is a way to tell CircleCI to provide secrets defined in
the Solidus Core Team context only to users that are part of
that group (on GitHub) when a build is triggered.

We only perform these Stoplight jobs in the master branch and
only the core team has the permissions to push to master.