Set payment method to none
By default Solidus is now looking for `solidus_paypal_commerce_platform`,
however, with the sandbox and the dummy generator we need to set this
to `none`, otherwise these scripts are not running successfully.

This commit fixes that. It does not update the application template for Heroku
though. That will be a seperate PR.