Make admin order event_links translatable
The admin/orders_helper event_links didn't had enough scope to be translatable for the particular order event. In some languages canceling and order is something different than clicking a "cancel" button in a form.

The solidus_i18n gem already has the translation scopes for these events, so I implemeted it that it will Just Work™