Migrate default billing addresses to address book
`Spree::User` records have a `bill_address_id` that denotes their saved
billing address to be reused in the next checkout. That column, however,
is not in use since Solidus 2.11.0. Instead, we rely on a
`default_billing` Boolean on the `spree_user_addresses` join table.

This migration sets the `default_billing` flag on that join table for
addresses matching the user's default billing address and the user's ID.

MySQL has different SQL syntax than standard SQL, which is why we need
two SQL statements.

For stores that have run this or a similar migration before, there's a
safeguard so they remember to fix their migrations when they upgrade.