Migrate default billing addresses to address book
`Spree::User` records have a `bill_address_id` that denotes their saved
billing address to be reused in the next checkout. That column, however,
is not in use since Solidus 2.11.0. Instead, we rely on a`default_billing`
Boolean on the `spree_user_addresses` join table.

This migration sets the `default_billing` flag on that join table for
addresses matching the user's default billing address and the user's ID.

For stores that have already started using the `default_billing`
Boolean, this task excludes the user_addresses that are already using
the new format by setting those `bill_address_id`s to `nil` before
running the query.

Co-authored-by: mamhoff <mamhoff@gmail.com>
Co-authored-by: DanielePalombo <danielepalombo@nebulab.it>