Fix preferences columns to be always serialized
This fixes an issue that arose because of commit b947015b, which
initialized the serialization of the preference column only if the
respective subclass had any preferences defined. This is not always the
case, and previous versions of Solidus would always allow a subclass of
e.g. `Spree::Calculator` to access their `preferences` attribute as a
Hash.

The intention of the commit in question was to speed up initialization
for models that do not have a preferences column. Instead of complex
metamagic, this moves the initialization of the preferences column and
preference defaults to those classes that actually needed, actually
yielding MORE performance benefits, as `serialize` and the
`after_initialize` hook are only called once, not once per preference
being defined.