Remove promotion_code#adjustments
This association is not used anywhere in the codebase, and it's probably
a remnant of adjustments being directly associated with promotion codes.

SolidusLegacyPromotions does this, but we have no need for this, because
we can just as well join via the `friendly_order_promotions` table.
An adjustment from a promotion with codes will have one of the promotion
benefits as a source, and an order_promotion with a reference to the
promotion code used.

I could have also changed the association to one that follows this path,
but I'm really not sure what we would use this association for.